refacto: controllers refacto
